th10: option to show real bullet sprites on spells with bullets that change type
This implements #63. Credit to CyrusVorazan for the Japanese
translation. I put that translation into DeepL for the Chinese text
because I got impatient and it looked similair to when I put an English
version into DeepL so I just went with it. Chinese people will probably
cringe at it though