Delete `dllmain.cpp`
The entire file is `#if 0`'d, and I can't find any references to it in the project files. And MSVC doesn't complain when I delete it, either...? 🤔

My best guess is that the file is left over from a previous version of thprac, and that perhaps Ack removed it from the MSVC project without deleting it outright?