feat: include a thprac98 release executables file
I think the main thprac executables should always stub thprac98 releases
that are *stable*, i.e., have a version number attached.

Somehow, the behavior of the stubbed main executables and the PC-98-only
executables is different. This will be fixed when stubbing the
thprac98 v0.0.1, which will be compiled using modern libraries and
compilers (and, hopefully, linkers and assemblers).

Due to the reason stated above, the linker option `/stub` is
temporarily removed.